+++ tests/CryptoExchanges.Net.Bybit.Tests.Unit/BybitMappingAndServiceTests.cs
@@ -11,7 +11,6 @@ using CryptoExchanges.Net.Core;
 using CryptoExchanges.Net.Core.Interfaces;
 using CryptoExchanges.Net.Core.Models;
 using CryptoExchanges.Net.Core.Enums;
-using CryptoExchanges.Net.DependencyInjection;
 
 namespace CryptoExchanges.Net.Bybit.Tests.Unit;
 
@@ -333,17 +332,6 @@ public class BybitMappingAndServiceTests
         services.AddBybitExchange(o => o.TimeoutSeconds = 0);
         var act = () => services.BuildServiceProvider().GetRequiredKeyedService<IExchangeClient>(ExchangeId.Bybit);
         act.Should().Throw<Microsoft.Extensions.Options.OptionsValidationException>();
-    }
-
-    [Fact]
-    public async Task Di_AddCryptoExchanges_ResolvesBybitAndBinance()
-    {
-        var services = new ServiceCollection();
-        services.AddCryptoExchanges();
-        await using var sp = services.BuildServiceProvider();
-
-        sp.GetRequiredKeyedService<IExchangeClient>(ExchangeId.Bybit).ExchangeId.Should().Be(ExchangeId.Bybit);
-        sp.GetRequiredKeyedService<IExchangeClient>(ExchangeId.Binance).ExchangeId.Should().Be(ExchangeId.Binance);
     }
 
     [Fact]